requirements.txt
flask==3.1.0
flask_cors==5.0.1
python-dotenv==1.0.1
requests==2.32.3
openai
